WJ-102 - Update: Move enum from utils to shared;
