+++ src/domains/Movie/api/Recommendations/index.ts
@@ -2,7 +2,7 @@ import tmdb from 'services/api/tmdb';
 
 import { formatDate, formatTmdbImage } from 'shared/utils';
 
-import { EntityType } from 'shared/utils/enums';
+import { EntityType } from 'shared/enums';
 import RawResponse from 'domains/Movie/api/Recommendations/RawResponse';
 import Response from 'domains/Movie/api/Recommendations/Response';
