+++ src/shared/enums/EntityType.ts
