+++ src/shared/enums/index.ts
@@ -1,4 +1,5 @@
 export { default as Color } from './Color';
+export { default as EntityType } from './EntityType';
 export { default as PosterHeight } from './PosterHeight';
 export { default as PosterWidth } from './PosterWidth';
 export { default as Size } from './Size';
