+++ src/shared/utils/enums/index.ts
@@ -1,2 +1,1 @@
-export { default as EntityType } from './EntityType';
 export { default as TmdbImageSize } from './TmdbImageSize';
